Addressed David's comment about restructuring
