+++ frontend/src/ScssComponents/global.scss
@@ -1,0 +1,2 @@
+@import './root.scss';
+@import '../components/Header/Header.scss';
