--- frontend/src/scss-components/global.scss
@@ -1,2 +1,0 @@
-@import './root.scss';
-@import './header.scss';
